feat(vsrc): add hex display and shift register modules
Modify `segments_display` to correct segment representation for number 9 and add support for hexadecimal numbers.

Add `shift_register` and `top_shift_register` modules to implement an 8-bit linear-feedback shift register (LFSR) as a simple pseudo-random number generator. Connect the output of the LFSR to 7-segment displays to show the generated number in hexadecimal format.

Test the functionality on the DE10-Standard development board.